+++ packages/core/test/one/index.js
@@ -24,9 +24,9 @@ export const hooks = {
     O.foo *= 2;
   },
   '@flecks/core/test/invoke-merge': () => ({foo: 69}),
-  '@flecks/core/test/invoke-merge-async': () => new Promise((resolve) => resolve({foo: 69})),
+  '@flecks/core/test/invoke-merge-async': () => new Promise((resolve) => { resolve({foo: 69}); }),
   '@flecks/core/test/invoke-merge-unique': () => ({foo: 69}),
-  '@flecks/core/test/invoke-merge-unique-async': () => new Promise((resolve) => resolve({foo: 69})),
+  '@flecks/core/test/invoke-merge-unique-async': () => new Promise((resolve) => { resolve({foo: 69}); }),
   '@flecks/core/test.middleware': () => (foo, next) => {
     // eslint-disable-next-line no-param-reassign
     foo.bar += 1;