+++ packages/repl/src/repl.js
@@ -62,7 +62,9 @@ export async function createReplServer(flecks) {
   }
   const socket = join(tmpdir(), 'flecks', id, 'repl', `${id}-${Date.now()}.sock`);
   flecks.set('$flecks/repl.socket', socket);
-  await new Promise((resolve) => netServer.listen(socket, resolve));
+  await new Promise((resolve) => {
+    netServer.listen(socket, resolve);
+  });
   debug('listening @ %s', socket);
 }