+++ packages/socket/src/packet/index.js
@@ -1,3 +1,3 @@
 export {default as Packer} from './packer';
-export {default} from './packet';
+export {default as Packet} from './packet';
 export {default as ValidationError} from './validation-error';